将Long对应的JDBC type改为BIGINT

+++ src/main/java/com/dormitory/dao/CampusCardDAO.java
@@ -15,7 +15,7 @@ public interface CampusCardDAO {
 	@Select("select * from campus_card c where c.student_id=#{0}")
 	@Results(value = {
             @Result(id = true, property = "campusCardId", column = "campus_card_id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
-            @Result(property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.DECIMAL),
+            @Result(property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
             @Result(property = "cardBalance", column = "card_balance", javaType = BigDecimal.class, jdbcType = JdbcType.DECIMAL),
             @Result(property = "hotwaterBalance", column = "hotwater_balance", javaType = BigDecimal.class, jdbcType = JdbcType.DECIMAL)
             })
