+++ src/main/java/com/dormitory/dao/NetworkDAO.java
@@ -14,7 +14,7 @@ public interface NetworkDAO {
    
 	@Select("select * from network n  where n.student_id=#{0} ")
 	@Results(value = {
-	        @Result(id = true, property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.DECIMAL),
+	        @Result(id = true, property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
 	        @Result(property = "tariffPackage", column = "tariff_package", javaType = String.class, jdbcType = JdbcType.VARCHAR),
 	        @Result(property = "money", column = "money", javaType = Integer.class, jdbcType = JdbcType.INTEGER)
 	})
