+++ src/main/java/com/dormitory/dao/NetworkRecordDAO.java
@@ -18,7 +18,7 @@ public interface NetworkRecordDAO {
 			+ " where  r.student_id=#{0}")
 	@Results(value = {
 			@Result(id = true, property = "networkRecordId", column = "network_record_id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
-			@Result(property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.DECIMAL),
+			@Result(property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
 			@Result(property = "tariffPackage", column = "tariff_package", javaType = String.class, jdbcType = JdbcType.VARCHAR),
 			@Result(property = "money", column = "money", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
 			@Result(property = "createTime", column = "create_time", javaType = Date.class, jdbcType = JdbcType.DATE),
