+++ src/main/java/com/dormitory/dao/StudentDAO.java
@@ -12,7 +12,7 @@ public interface StudentDAO {
     
 	@Select("select * from student where student_id=#{0}")
 	@Results(value = {
-            @Result(id = true, property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.DECIMAL),
+            @Result(id = true, property = "studentId", column = "student_id", javaType = Long.class, jdbcType = JdbcType.BIGINT),
             @Result( property = "dormitoryId", column = "dormitory_id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
             @Result( property = "bedId", column = "bed_id", javaType = Integer.class, jdbcType = JdbcType.INTEGER),
             @Result(property = "name", column = "name", javaType = String.class, jdbcType = JdbcType.VARCHAR),
